requirements.txt
fastapi 
requests
bs4
duckduckgo-search
ddgs
spacy
uvicorn
twelvelabs
requests
yt-dlp
typing
instaloader
ffmpeg
google-genai
fastapi
uvicorn
httpx
python-dotenv
pydantic
httpx
python-dotenv
pydantic
pytest
